move telemetry tests to dedicated unit test file
